+++ packages/saas-ui-forms/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @saas-ui/forms
 
+## 3.0.0-alpha.22
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @saas-ui/react@3.0.0-alpha.22
+
 ## 3.0.0-alpha.21
 
 ### Patch Changes